requirements.txt
selenium
requests
python-dotenv
selenium_stealth
chromedriver_autoinstaller